integrating c++17 fixes from #959 (ref:https://github.com/flann-lib/flann/pull/392), fixed pop_t error (https://github.com/flann-lib/flann/commit/b6e37553de0dd47dc8fc9035ba5aef163981b6a4)

+++ corelib/include/rtabmap/core/FlannIndex.h
@@ -41,10 +41,10 @@ public:
 	virtual ~FlannIndex();
 
 	void release();
-	unsigned int indexedFeatures() const;
+	size_t indexedFeatures() const;
 
 	// return Bytes
-	unsigned long memoryUsed() const;
+	size_t memoryUsed() const;
 
 	// Note that useDistanceL1 doesn't have any effect if LSH is used
 	void buildLinearIndex(
